+++ src/main/java/com/web4enterprise/pdf/core/page/Page.java
@@ -138,7 +138,7 @@ public class Page implements PdfObject, PageNode {
 	 * @param size The size of the text to add.
 	 * @param text The text to add.
 	 */
-	public void addText(float x, float y, int size, String text) {
+	public void addText(float x, float y, float size, String text) {
 		add(new Text(x, y, size, text));
 	}
 	
@@ -151,7 +151,7 @@ public class Page implements PdfObject, PageNode {
 	 * @param fontVariant The font variant of the text.
 	 * @param text The text to add.
 	 */
-	public void addText(float x, float y, int size, FontVariant fontVariant, String text) {
+	public void addText(float x, float y, float size, FontVariant fontVariant, String text) {
 		add(new Text(x, y, size, fontVariant, text));
 	}
 	
@@ -165,7 +165,7 @@ public class Page implements PdfObject, PageNode {
 	 * @param color The color of the text.
 	 * @param text The text to add.
 	 */
-	public void addText(float x, float y, int size, FontVariant fontVariant, Color color, String text) {
+	public void addText(float x, float y, float size, FontVariant fontVariant, Color color, String text) {
 		add(new Text(x, y, size, fontVariant, color, text));
 	}
